// packages/types/src/generated/prisma/models/work_item_worklogs.ts
export type work_item_worklogsSelect<
  ExtArgs extends runtime.Types.Extensions.InternalArgs =
    runtime.Types.Extensions.DefaultArgs,
> = runtime.Types.Extensions.GetSelect<
  {
    id?: boolean;
    work_item_id?: boolean;
    user_id?: boolean;
    logged_hours?: boolean;
    logged_at?: boolean;
    comment?: boolean;
    status?: boolean;
    created_by?: boolean;
    created_at?: boolean;
    updated_by?: boolean;
    updated_at?: boolean;
    work_item?: boolean | Prisma.work_itemsDefaultArgs<ExtArgs>;
    user?: boolean | Prisma.usersDefaultArgs<ExtArgs>;
  },
  ExtArgs['result']['work_item_worklogs']
>;
// ...
export type work_item_worklogsOmit<
  ExtArgs extends runtime.Types.Extensions.InternalArgs =
    runtime.Types.Extensions.DefaultArgs,
> = runtime.Types.Extensions.GetOmit<
  | 'id'
  | 'work_item_id'
  | 'user_id'
  | 'logged_hours'
  | 'logged_at'
  | 'comment'
  | 'status'
  | 'created_by'
  | 'created_at'
  | 'updated_by'
  | 'updated_at',
  ExtArgs['result']['work_item_worklogs']
>;
export type work_item_worklogsInclude<
  ExtArgs extends runtime.Types.Extensions.InternalArgs =
    runtime.Types.Extensions.DefaultArgs,
> = {
  work_item?: boolean | Prisma.work_itemsDefaultArgs<ExtArgs>;
  user?: boolean | Prisma.usersDefaultArgs<ExtArgs>;
};
// ...
export type $work_item_worklogsPayload<
  ExtArgs extends runtime.Types.Extensions.InternalArgs =
    runtime.Types.Extensions.DefaultArgs,
> = {
  name: 'work_item_worklogs';
  objects: {
    work_item: Prisma.$work_itemsPayload<ExtArgs>;
    user: Prisma.$usersPayload<ExtArgs>;
  };
  scalars: runtime.Types.Extensions.GetPayloadResult<
    {
      id: string;
      work_item_id: string;
      user_id: string;
      logged_hours: number;
      logged_at: Date;
      comment: string | null;
      status: $Enums.RecordStatus;
      created_by: string | null;
      created_at: Date;
      updated_by: string | null;
      updated_at: Date;
    },
    ExtArgs['result']['work_item_worklogs']
  >;
  composites: {};
};

export type work_item_worklogsGetPayload<
  S extends boolean | null | undefined | work_item_worklogsDefaultArgs,
> = runtime.Types.Result.GetResult<Prisma.$work_item_worklogsPayload, S>;
// ...
export type work_item_worklogsDefaultArgs<
  ExtArgs extends runtime.Types.Extensions.InternalArgs =
    runtime.Types.Extensions.DefaultArgs,
> = {
  /**
   * Select specific fields to fetch from the work_item_worklogs
   */
  select?: Prisma.work_item_worklogsSelect<ExtArgs> | null;
  /**
   * Omit specific fields from the work_item_worklogs
   */
  omit?: Prisma.work_item_worklogsOmit<ExtArgs> | null;
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.work_item_worklogsInclude<ExtArgs> | null;
};